+++ GamePlan.Web/src/Components/Week.jsx
@@ -7,7 +7,23 @@ const Week = () => {
       const date = new Date();
       date.setDate(date.getDate() - date.getDay() + i + 1);
       weekDays.push(date);
-    }
+  }
+  
+  const getWeekNumber = (d) => {
+    d = new Date(Date.UTC(d.getFullYear(), d.getMonth(), d.getDate()));
+    d.setUTCDate(d.getUTCDate() + 4 - (d.getUTCDay() || 7));
+    var yearStart = new Date(Date.UTC(d.getUTCFullYear(), 0, 1));
+    var weekNo = Math.ceil(((d - yearStart) / 86400000 + 1) / 7);
+    return [d.getFullYear(), weekNo];
+  }
+
+  const handleNextWeek = () => {
+    setWeekNumber(prevWeek => prevWeek + 1);
+  };
+
+  const handlePrevWeek = () => {
+    setWeekNumber(prevWeek => prevWeek - 1);
+  };
 
   return (
     <>
